Break circular dependencies
