+++ ts/libs/core/src/lib/composed-metrics/public/common/index.ts
@@ -1,4 +1,3 @@
-export * from './composed-metric-params';
 export * from './composed-metric-source-factory';
 export * from './composed-metric-source.base';
 export * from './composed-metric-source';
