+++ ts/libs/core/src/lib/composed-metrics/public/control/composed-metric-collector.ts
@@ -1,5 +1,6 @@
+import { ComposedMetricParams } from '../../../model';
 import { Sample } from '../../../raw-metrics-query';
-import { ComposedMetricParams, ComposedMetricType } from '../common';
+import { ComposedMetricType } from '../common';
 
 /**
  * Used to collect samples of a composed metric, which are then sent to a DB or cached for
