+++ ts/libs/core/src/lib/composed-metrics/public/control/default-composed-metric-mapping-manager.ts
@@ -1,7 +1,8 @@
-import { ApiObjectMetadata, ComposedMetricMapping, ComposedMetricMappingSpec, ObjectKind, OwnerReference } from '../../../model';
+import { pascalCase } from 'change-case';
+import { ApiObjectMetadata, ComposedMetricMapping, ComposedMetricMappingSpec, ComposedMetricParams, ObjectKind, OwnerReference } from '../../../model';
 import { OrchestratorClient } from '../../../runtime';
 import { Logger, POLARIS_API, isValueEqual } from '../../../util';
-import { ComposedMetricParams, ComposedMetricType } from '../common';
+import { ComposedMetricError, ComposedMetricType } from '../common';
 import { ComposedMetricMappingManager } from './composed-metric-mapping-manager';
 
 /**
@@ -15,7 +16,7 @@ export class DefaultComposedMetricMappingManager implements ComposedMetricMappin
         metricType: ComposedMetricType<V, P>,
         params: P,
     ): Promise<ComposedMetricMapping> {
-        const kind = ComposedMetricMapping.getMappingObjectKind(metricType);
+        const kind = this.getMappingObjectKind(metricType);
 
         let mapping = await this.fetchMetricMapping(kind, params);
         if (!mapping) {
@@ -29,6 +30,20 @@ export class DefaultComposedMetricMappingManager implements ComposedMetricMappin
         }
 
         return mapping;
+    }
+
+    getMappingObjectKind(metricType: ComposedMetricType<any>, override?: Partial<ObjectKind>): ObjectKind {
+        const gvkComponents = metricType.metricTypeName.split('/');
+        if (gvkComponents.length !== 3) {
+            throw new ComposedMetricError('Metric type does not conform to the `group/version/kind` naming standard.', metricType);
+        }
+
+        const gvk = new ObjectKind({
+            group: override?.group ?? gvkComponents[0],
+            version: override?.version ?? gvkComponents[1],
+            kind: override?.kind ?? pascalCase(gvkComponents[2]) + 'MetricMapping',
+        });
+        return gvk;
     }
 
     private async fetchMetricMapping(kind: ObjectKind, params: ComposedMetricParams): Promise<ComposedMetricMapping | undefined> {
