+++ ts/libs/core/src/lib/model/index.ts
@@ -1,11 +1,13 @@
-export * from './api-object.prm';
 export * from './api-object-metadata.prm';
+export * from './api-object.prm';
 export * from './common-elasticity-configs.prm';
 export * from './composed-metric-mapping.prm';
+export * from './composed-metric-params.prm';
 export * from './container.prm';
 export * from './custom-resource-definition.prm';
-export * from './elasticity-strategy.prm';
 export * from './elasticity-strategy-kind.prm';
+export * from './elasticity-strategy.prm';
+export * from './model-utils';
 export * from './object-kind.prm';
 export * from './object-reference.prm';
 export * from './owner-reference.prm';
