+++ ts/libs/core/src/lib/runtime/internal/index.ts
@@ -1,2 +1,1 @@
 export * from './metrics-source';
-export * from './polaris-microcontroller-factory-metadata';
