+++ ts/libs/core/src/lib/runtime/public/microcontroller-factory/default-microcontroller-factory.ts
@@ -1,7 +1,8 @@
-import { Constructor, PolarisMetadataUtils } from '../../../util';
-import { PolarisMicrocontrollerFactoryMetadata } from '../../internal';
+import { PolarisMicrocontrollerFactoryMetadata } from '../../../transformation/internal/polaris-microcontroller-factory-metadata';
+import { PolarisMetadataUtils } from '../../../transformation/internal/reflect-metadata-utils';
+import { Constructor, FactoryFn } from '../../../util';
 import { MicrocontrollerFactoryNotRegisteredError } from './errors';
-import { FactoryFn, MicrocontrollerFactory } from './microcontroller-factory';
+import { MicrocontrollerFactory } from './microcontroller-factory';
 
 export class DefaultMicrocontrollerFactory<S, C> implements MicrocontrollerFactory<S, C> {
