+++ ts/libs/core/src/lib/transformation/internal/index.ts
@@ -1,2 +1,4 @@
+export * from './polaris-microcontroller-factory-metadata';
 export * from './polaris-transformation-metadata';
 export * from './property-transformer';
+export * from './reflect-metadata-utils';
