+++ ts/libs/core/src/lib/transformation/public/common/index.ts
@@ -1,4 +1,6 @@
 export * from './errors';
 export * from './polaris-transformation-config';
+export * from './polaris-transformation-service-manager';
+export * from './polaris-transformation-service';
 export * from './polaris-transformer';
 export * from './polaris-type';
