+++ ts/libs/core/src/lib/transformation/public/service/index.ts
@@ -1,2 +1,1 @@
 export * from './default-polaris-transformation.service';
-export * from './polaris-transformation-service';
