+++ ts/libs/core/src/lib/transformation/public/transformers/default-transformer.ts
@@ -2,8 +2,7 @@ import { classToPlain, plainToClass } from 'class-transformer';
 import { cloneDeep } from 'lodash';
 import { JsonSchema } from '../../../model';
 import { Constructor, InterfaceOf, cloneDeepWithoutExcluded } from '../../../util';
-import { PolarisTransformer } from '../common';
-import { PolarisTransformationService } from '../service';
+import { PolarisTransformationService, PolarisTransformer } from '../common';
 
 /**
  * This transformer does not alter the structure of the objects, it just performs a simple
