+++ ts/libs/core/src/lib/util/index.ts
@@ -1,2 +1,1 @@
-export * from './internal';
 export * from './public';
