--- ts/libs/core/src/lib/util/internal/index.ts
@@ -1,1 +1,0 @@
-export * from './reflect-metadata-utils';
