+++ ts/libs/core/src/lib/util/public/index.ts
@@ -4,6 +4,5 @@ export * from './logger';
 export * from './model-utils';
 export * from './object-utils';
 export * from './observable-stopper';
-export * from './owner-utils';
 export * from './string-utils';
 export * from './util-types';
